+++ entity/user.go
@@ -3,7 +3,7 @@ package entity
 import "time"
 
 type User struct {
-	ID         int64        `json:"id" gorm:"primaryKey"`
+	ID         int          `json:"id" gorm:"primaryKey"`
 	Name       string       `json:"name"`
 	Email      string       `json:"email"`
 	CreatedAt  time.Time    `json:"created_at"`
