+++ repository/submissionRepository.go
@@ -8,7 +8,7 @@ import (
 
 type SubmissionRepository interface {
 	GetAll(ctx context.Context) ([]*entity.Submission, error)
-	GetById(ctx context.Context, id int64) (*entity.Submission, error)
+	GetById(ctx context.Context, id int) (*entity.Submission, error)
 	Create(ctx context.Context, submission *entity.Submission) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id int) error
 }
