+++ repository/submissionRepositoryImpl.go
@@ -30,7 +30,7 @@ func (r *SubmissionRepositoryImpl) GetAll(ctx context.Context) ([]*entity.Submis
 	return submissions, nil
 }
 
-func (r *SubmissionRepositoryImpl) GetById(ctx context.Context, id int64) (*entity.Submission, error) {
+func (r *SubmissionRepositoryImpl) GetById(ctx context.Context, id int) (*entity.Submission, error) {
 	var submission *entity.Submission
 	
 	if err := r.DB.Where("id = ?", id).First(&submission).Error; err != nil {
@@ -40,6 +40,6 @@ func (r *SubmissionRepositoryImpl) GetById(ctx context.Context, id int64) (*enti
 	return submission, nil
 }
 
-func (r *SubmissionRepositoryImpl) Delete(ctx context.Context, id int64) error {
+func (r *SubmissionRepositoryImpl) Delete(ctx context.Context, id int) error {
 	return r.DB.Where("id = ?", id).Delete(&entity.Submission{}).Error
 }
